Reorganizing app to isolate responsibilities. Adding composition folder to separate from other architectural folders
